Modificaciones finales en clase 08-10-20 astah y codigo

+++ Dominio/Excursion.cs
@@ -70,7 +70,7 @@ namespace Dominio
         public override string ToString()
         {
             //calculo costo total
-            int costoTotal=0;
+            double costoTotal=0;
             string destinosDisponibles = "";
             foreach (Destino i in this.ListaDestinosDisponibles)
             {
